+++ tests/test_active_power.py
@@ -13,8 +13,8 @@ from psdm.steadystate_case.controller import ControlPConst
 from psdm.steadystate_case.controller import ControlQConst
 from psdm.steadystate_case.controller import PController
 from psdm.steadystate_case.controller import QController
-from psdm.topology.load import ActivePower as ActivePowerSet
-from psdm.topology.load import ReactivePower as ReactivePowerSet
+from psdm.topology.quantities import ActivePower as ActivePowerSet
+from psdm.topology.quantities import ReactivePower as ReactivePowerSet
 
 
 class TestActivePower:
