+++ tests/test_branch.py
@@ -5,6 +5,10 @@
 from psdm.base import VoltageSystemType
 from psdm.topology.branch import Branch
 from psdm.topology.branch import BranchType
+from psdm.topology.quantities import Admittance
+from psdm.topology.quantities import Current
+from psdm.topology.quantities import Impedance
+from psdm.topology.quantities import Voltage
 
 
 class TestBranch:
@@ -13,12 +17,12 @@ class TestBranch:
             node_1="asd",
             node_2="fgh",
             name="wqertasd",
-            u_n=1,
-            i_r=1,
-            b1=1,
-            g1=1,
-            x1=1,
-            r1=1,
+            u_n=Voltage(values=(1,)),
+            i_r=Current(values=(1,)),
+            b1=Admittance(value=1),
+            g1=Admittance(value=1),
+            x1=Impedance(value=1),
+            r1=Impedance(value=1),
             type=BranchType.LINE,
             voltage_system_type=VoltageSystemType.AC,
         )
