+++ tests/test_reactive_power.py
@@ -12,8 +12,8 @@ from psdm.steadystate_case.controller import ControlQConst
 from psdm.steadystate_case.controller import PController
 from psdm.steadystate_case.controller import QController
 from psdm.steadystate_case.reactive_power import ReactivePower
-from psdm.topology.load import ActivePower as ActivePowerSet
-from psdm.topology.load import ReactivePower as ReactivePowerSet
+from psdm.topology.quantities import ActivePower as ActivePowerSet
+from psdm.topology.quantities import ReactivePower as ReactivePowerSet
 
 
 class TestReactivePower:
